+++ CHANGES.rst
@@ -1,6 +1,29 @@
 Change Log
 ----------
 
+1.1.11
+~~~~~~
+
+Released on December 5, 2012.
+
+* Fix bugs in the search and add device forms
+
+1.1.10
+~~~~~~
+
+Released on December 5, 2012.
+
+* support for SNMPv3 in discovery
+
+* DHCP config improvements: proper hostnames from PTR records; support for
+  syncing entries and networks from a specific DC only
+
+* DNS/DHCP addresses tab redesigned for usability and performance
+
+* improved search for software components and discovering software versions
+
+* discovery fixes
+
 1.1.9
 ~~~~~